Removed Dialog-Popups completely. Added Window-Title Bar Navigation: - Click on Settings will add the tabitem with its content. Clicking on Settings again will close the window
Added Projekt Management functionality:
- You can add, edit and delete projects
=> Name / Fullpath
It lets you create in the subfolder "projects" a project folder for future use for translations.
- Same clicking mechanism like Settings

You can select from the title bar the current project, clicking on it lets it add to the tabs.
This item is from type "ProjectWorkspace" It enables you to browse through the directories and files which you have configured via rootpath.

Made a fix for the encoding tables where \r \n were not correctly translated.

Optimizations ...